Code refactoring
Simplified __init__ method in Interface() by exporting some functionality to dedicated methods in the class.
This methods can now be called separately during unit testing (instead of the whole __init__ method, which does a bunch of stuff and thus is not good for doing UNIT testing)
Change the __main__ section of arduino and camera modules to not load Interface but just an instance of their classes.