Renaming and improved program writing
Now user only needs to supply a short program file that will contain blocks
Each block has a name, start time (minutes), duration (minutes) and times to be run
The name of the block should be equal to the filename of one of the files (without extension) in the blocks/ directory (the block file, containing pin events).

Behind the scenes, the program will add the start of each block to the start and end the events in the block file. Moreover, if times is n, all the blocks get copied n times and their start and end get added (duration * (times -1)).

This makes program writing very easy and modular, where complex programs can be built by making use of blocks which may already exist in the blocks/ dir and can anyway be newly specified